Refactor

+++ app/simulationcontroller.cpp
@@ -25,7 +25,7 @@ SimulationController::SimulationController(Scene *scn)
     , m_shouldRestart(false)
 {
     m_simulationTimer.setInterval(globalClock);
-    m_viewTimer.setInterval(int(1000 / 30));
+    m_viewTimer.setInterval(1000 / 30);
     m_viewTimer.start();
     connect(&m_viewTimer, &QTimer::timeout, this, &SimulationController::updateView);
     connect(&m_simulationTimer, &QTimer::timeout, this, &SimulationController::tic);
